Refactor module structures and update import paths
Reorganized several modules, mainly moving files from the lib directory to the cli directory.
