+++ foundrytools_cli_2/cli/converter/cli.py
@@ -18,8 +18,8 @@ from foundrytools_cli_2.cli.shared_options import (
     subroutinize_flag,
     target_upm_option,
 )
+from foundrytools_cli_2.cli.task_runner import TaskRunner
 from foundrytools_cli_2.lib.font import Font
-from foundrytools_cli_2.lib.task_runner import TaskRunner
 
 cli = click.Group("converter", help="Font conversion utilities.")
