+++ foundrytools_cli_2/cli/converter/snippets/sfnt_to_web.py
@@ -1,9 +1,9 @@
 import typing as t
 from pathlib import Path
 
+from foundrytools_cli_2.cli.logger import logger
 from foundrytools_cli_2.lib.constants import WOFF2_FLAVOR, WOFF_FLAVOR
 from foundrytools_cli_2.lib.font import Font
-from foundrytools_cli_2.lib.logger import logger
 
 
 def main(
