+++ foundrytools_cli_2/cli/converter/snippets/ttf_to_otf.py
@@ -3,9 +3,9 @@ from pathlib import Path
 
 from afdko.fdkutils import run_shell_command
 
+from foundrytools_cli_2.cli.logger import logger
 from foundrytools_cli_2.lib.font import Font
 from foundrytools_cli_2.lib.font.tables import OS2Table
-from foundrytools_cli_2.lib.logger import logger
 from foundrytools_cli_2.lib.otf.otf_builder import build_otf
 from foundrytools_cli_2.lib.otf.t2_charstrings import get_t2_charstrings
