+++ foundrytools_cli_2/cli/font_finder.py
@@ -4,8 +4,20 @@ from pathlib import Path
 
 from fontTools.ttLib.ttFont import TTLibError
 
-from foundrytools_cli_2.lib.constants import FinderOptions
 from foundrytools_cli_2.lib.font import Font
+
+__all__ = ["FinderError", "FinderFilter", "FinderOptions", "FontFinder"]
+
+@dataclass
+class FinderOptions:
+    """
+    A class that specifies the options to pass to the FontFinder class.
+    """
+
+    recursive: bool = False
+    lazy: t.Optional[bool] = None
+    recalc_bboxes: bool = True
+    recalc_timestamp: bool = False
 
 
 @dataclass
@@ -191,6 +203,3 @@ def _is_static(font: Font) -> bool:
 
 def _is_variable(font: Font) -> bool:
     return font.is_variable
-
-
-__all__ = ["FontFinder", "FinderError", "FinderFilter"]
