+++ foundrytools_cli_2/cli/otf/snippets/fix_contours.py
@@ -1,5 +1,5 @@
+from foundrytools_cli_2.cli.logger import logger
 from foundrytools_cli_2.lib.font import Font
-from foundrytools_cli_2.lib.logger import logger
 
 
 def main(font: Font, min_area: int = 25, subroutinize: bool = True) -> None:
