+++ foundrytools_cli_2/cli/post/cli.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import click
 
 from foundrytools_cli_2.cli.shared_options import base_options
-from foundrytools_cli_2.lib.task_runner import TaskRunner
+from foundrytools_cli_2.cli.task_runner import TaskRunner
 
 
 @click.command(no_args_is_help=True)
