--- foundrytools_cli_2/lib/__init__.py
@@ -1,11 +1,0 @@
-from foundrytools_cli_2.lib.font_finder import FontFinder
-from foundrytools_cli_2.lib.logger import logger
-from foundrytools_cli_2.lib.task_runner import TaskRunner
-from foundrytools_cli_2.lib.timer import Timer
-
-__all__ = [
-    "FontFinder",
-    "TaskRunner",
-    "logger",
-    "Timer",
-]
